Spatial Memory Model of Mice
This program models the research work by Richard G.M. Morris titled
"Spatial localization does not require the presence of local cues". In
this program a mouse (M) learns the location of a hidden platform (P),
through a series of trials. While the platform is visible throughout
the program to the user, the mouse is unaware of the platform's
location and must learn the platform's location through a series of
trials where the mouse searches the watermaze for the platform. The
mouse can only tell if it found the platform if it is at the platform's
location. In the beginning, the mouse searches randomly for the
platform, but after each trial, the mouse begins to move more directly
towards the platform. How quickly the mouse learns the location of the
platform is significantly influenced by the value inputted to the
program at the start of the program. This value, which can be from 0 to
100% indicates the quality of the mouse's spatial memory.